tests for rest and mariadb connection with testEntity

+++ src/main/java/com/schenkbarnabas/tlog16rs/core/beans/WorkDayRB.java
@@ -10,5 +10,5 @@ public class WorkDayRB {
     private int year;
     private int month;
     private int day;
-    private int requiredHours;
+    private double requiredHours;
 }
